Spoiler FCS improvement
Implement a lateral stick movement simulation.
Firstly to adjust the stick position to take into account the force that would be required. As we don't have any way of loading the controls (i.e. statically sprung joysticks at best) the only viable option is to convert the force into an equivalent position; i.e. that you will have to move the stick more  to achieve the same deflection.

The second part will then apply the spoiler deflection schedule resulting in a commanded spoiler position

This is based on  AP 101B-3101-1EE 19-20 Fig.10 and Fig. 12