Revised aero based on new geometry
This model uses revised geometry based on more a rescale and remeasure of the 3view diagrams that I took from Jane's together with improved reference point (AeroRP) and engine position.

Overall the roll rate is much improved, this is due to FCS improvements and a smaller aero model for the spoiler. There is much less adverse yaw and the aircraft is much more stable in pitch.