Pre 2019.1 JSBSim compatibility
Rework of the FCS to be compatible with 2018.3 (tested) and probably earlier versions.

With this version of JSBSim the systems can't reliably rebind or modify (via a name=) properties that are tied; which causes the 'exported' surface positions to be stuck at zero.

Remove the direct export to properties in /surface-positions - as this isn't necessary for standard properties.